flyttet tekst og ikon, fikset på layout i joyride-modalen, neste-knappen er bredere, endret tekst fra avslutt veiviser til avslutt
