+++ frontend/mulighetsrommet-veileder-flate/src/components/joyride/OversiktenJoyride.tsx
@@ -1,6 +1,6 @@
 import Joyride, { ACTIONS, CallBackProps, EVENTS, STATUS } from 'react-joyride';
 import { useEffect, useState } from 'react';
-import { joyrideKnappefarge, localeStrings } from './utils';
+import { joyrideStyling, localeStrings } from './Utils';
 import { JoyrideKnapp } from './JoyrideKnapp';
 import { logEvent } from '../../core/api/logger';
 import { getStepIndex, isStep, stepsOversikten } from './Steps';
@@ -110,11 +110,7 @@ export function OversiktenJoyride({ setHistorikkModalOpen, isHistorikkModalOpen 
         showSkipButton
         stepIndex={state.stepIndex}
         disableScrolling
-        styles={{
-          options: {
-            primaryColor: joyrideKnappefarge,
-          },
-        }}
+        styles={joyrideStyling()}
         disableCloseOnEsc={false}
         disableOverlayClose={true}
       />
