+++ frontend/mulighetsrommet-veileder-flate/src/components/joyride/Utils.ts
@@ -1,0 +1,21 @@
+export const localeStrings = () => ({
+  next: 'Neste',
+  back: 'Forrige',
+  close: 'Lukk',
+  skip: 'Avslutt',
+  last: 'Ferdig',
+});
+
+export const joyrideStyling = () => {
+  return {
+    options: {
+      primaryColor: '#0067C5',
+    },
+    tooltip: {
+      padding: 25,
+    },
+    buttonNext: {
+      width: 100,
+    },
+  };
+};
