--- frontend/mulighetsrommet-veileder-flate/src/components/joyride/utils.ts
@@ -1,9 +1,0 @@
-export const localeStrings = () => ({
-  next: 'Neste',
-  back: 'Forrige',
-  close: 'Lukk',
-  skip: 'Avslutt veiviser',
-  last: 'Ferdig',
-});
-
-export const joyrideKnappefarge = '#0067C5';
