+++ frontend/mulighetsrommet-veileder-flate/src/views/tiltaksgjennomforing-oversikt/ViewTiltaksgjennomforingOversikt.module.scss
@@ -14,21 +14,5 @@
 
   .knapperad {
     display: flex;
-    gap: 0.5rem;
-
-    button {
-      width: 5rem;
-      height: 4rem;
-
-      :global(.navds-label) {
-        display: flex;
-        flex-direction: column;
-        font-weight: lighter;
-
-        svg {
-          margin: 0 auto;
-        }
-      }
-    }
   }
 }
